add admin rules to server side

+++ server/controllers/UserController.js
@@ -1,22 +1,22 @@
 const User = require('../db/models/UserModel');
 const bcrypt = require('bcryptjs');
+
 // Create a new user in the database
 async function createUser(email, password, firstName, lastName, gender) {
-    // Check if user already exists in the database
-    let user = await User.findOne({ email });
-    if (user) {
-      throw new Error('User already exists');
-    }
-  
-    // Hash the password and create a new user in the database
-    const salt = await bcrypt.genSalt(10);
-    const hashedPassword = await bcrypt.hash(password, salt);
-    user = new User({ email, password: hashedPassword, firstName, lastName, gender });
-    await user.save();
-    console.log(`New user registered: ${firstName} ${lastName} ,email: (${email})`);
-    return user;
+  // Check if user already exists in the database
+  let user = await User.findOne({ email });
+  if (user) {
+    throw new Error('User already exists');
   }
-  
+
+  // Hash the password and create a new user in the database
+  const salt = await bcrypt.genSalt(10);
+  const hashedPassword = await bcrypt.hash(password, salt);
+  user = new User({ email, password: hashedPassword, firstName, lastName, gender });
+  await user.save();
+  console.log(`New user registered: ${firstName} ${lastName} ,email: (${email})`);
+  return user;
+}
 
 // Retrieve an existing user from the database
 async function getUser(id) {
@@ -29,14 +29,27 @@ async function getUser(id) {
 }
 
 // Update an existing user in the database
-async function updateUser(id, data) {
+async function updateUser(id, data, isAdmin = false) {
   const user = await getUser(id);
+
+  // Only allow admins to update users other than themselves
+  if (!isAdmin && user.id !== id) {
+    throw new Error('You do not have permission to update this user');
+  }
+
   Object.assign(user, data);
   await user.save();
 }
 
 // Delete an existing user from the database
-async function deleteUser(id) {
+async function deleteUser(id, isAdmin = false) {
+  const user = await getUser(id);
+
+  // Only allow admins to delete users other than themselves
+  if (!isAdmin && user.id !== id) {
+    throw new Error('You do not have permission to delete this user');
+  }
+
   await User.findByIdAndDelete(id);
 }
 
@@ -48,11 +61,10 @@ async function getUserByEmail(email) {
   return user;
 }
 
-
 module.exports = {
   createUser,
   getUser,
   updateUser,
   deleteUser,
-  getUserByEmail
+  getUserByEmail,
 };
